+++ packages/yourloops/test/integration/mock/team.api.mock.ts
@@ -31,7 +31,7 @@ import { TeamMemberRole } from '../../../lib/team/models/enums/team-member-role.
 import { TeamType } from '../../../lib/team/models/enums/team-type.enum'
 import { HcpProfession } from '../../../lib/auth/models/enums/hcp-profession.enum'
 
-import { UserInvitationStatus } from '../../../lib/team/models/enums/user-invitation-status.enum'
+import { UserInviteStatus } from '../../../lib/team/models/enums/user-invite-status.enum'
 import { type Team } from '../../../lib/team'
 import { type Profile } from '../../../lib/auth/models/profile.model'
 import { DEFAULT_THRESHOLDS_IN_MGDL } from '../../../components/monitoring-alert/monitoring-alert.default'
@@ -99,7 +99,7 @@ export const buildTeamOne = (): Team => {
         privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
         termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true }
       } as Profile,
-      status: UserInvitationStatus.accepted,
+      status: UserInviteStatus.Accepted,
       email: 'ylp.ui.test.hcp.28@diabeloop.fr'
     }]
   }
@@ -126,7 +126,7 @@ export const buildTeamTwo = (): Team => {
         privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
         termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true }
       } as Profile,
-      status: UserInvitationStatus.accepted,
+      status: UserInviteStatus.Accepted,
       email: 'ylp.ui.test.hcp.28@diabeloop.fr'
     }],
     monitoringAlertsParameters
@@ -154,7 +154,7 @@ export const buildTeamThree = (): Team => {
         privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
         termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true }
       } as Profile,
-      status: UserInvitationStatus.accepted,
+      status: UserInviteStatus.Accepted,
       email: 'ylp.ui.test.hcp.28@diabeloop.fr'
     }, {
       userId: '4d9afc649ae4',
@@ -168,7 +168,7 @@ export const buildTeamThree = (): Team => {
         privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
         termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true }
       } as Profile,
-      status: UserInvitationStatus.accepted,
+      status: UserInviteStatus.Accepted,
       email: 'ylp.ui.test.hcp.28.1@diabeloop.fr'
     }],
     monitoringAlertsParameters
@@ -196,7 +196,7 @@ export const buildFiltersTeam = (): Team => {
         privacyPolicy: { acceptanceTimestamp: '2021-05-22', isAccepted: true },
         termsOfUse: { acceptanceTimestamp: '2021-05-22', isAccepted: true }
       } as Profile,
-      status: UserInvitationStatus.accepted,
+      status: UserInviteStatus.Accepted,
       email: 'filters@user.com'
     }]
   }